test truffle

+++ private_network/private_network.sh
@@ -9,24 +9,26 @@ function set_dir {
 }
 
 function gen_pnet {
-    if [ ! -d ./var/main ]; then
+    if [ ! -d ./var/main/geth ]; then
         geth --datadir ./var/main init genesis.json
+        echo "init private network"
     fi
 
     if [ -f .pnet ]; then
         running_pid=$(cat .pnet)
 
-        kill -9 $running_pid
+        kill -15 $running_pid
         rm .pnet
-        echo "kill already running pnet process: $running_pid"
+        echo "kill already running private network process: $running_pid"
     fi
 
     {
-        geth --mine --datadir ./var/main --networkid 333 2> ./var/log/pnet.log &
+        geth --unlock "0x6e1a9ec327f437fb1ef62900842e7ca1864580c3" --password ".password" --nodiscover --rpc --rpccorsdomain "http://localhost:8080" --mine --datadir ./var/main --networkid 333 2> ./var/log/pnet.log &
         pnet_pid=$! 
         echo $pnet_pid > .pnet
+        echo "start private network"
     } || {
-        kill -9 $pnet_pid
+        kill -15 $pnet_pid
         rm .pnet
         echo "error: generate private network: kill process: $pnet_pid"
     }
@@ -35,7 +37,7 @@ function gen_pnet {
 function gen_discover {
     if [ -f .bootnode ]; then
         running_pid=$(cat .bootnode)
-        kill -9 $running_pid
+        kill -15 $running_pid
         rm .bootnode
         echo "kill already running bootnode process: $running_pid"
     fi
@@ -44,8 +46,9 @@ function gen_discover {
         bootnode --nodekey=pnet.key 2> ./var/log/discover.log &
         bootnode_pid=$!
         echo $bootnode_pid > .bootnode
+        echo "start discovery"
     } || {
-        kill -9 $bootnode_pid
+        kill -15 $bootnode_pid
         rm .bootnode
         echo "error: generate bootnode: kill process: $bootnode_pid"
     }
@@ -66,30 +69,41 @@ function list_process {
 }
 
 function kill_process {
-    local pnet_pid=$(cat .pnet)
-    local bootnode_pid=$(cat .bootnode)
-    local node_pid=$(cat .node)
+    if [ -f .pnet ]; then
+        local pnet_pid=$(cat .pnet)
 
-    if [ ! -z $pnet_pid ]; then
-        kill -9 $pnet_pid
-        rm .pnet    
+        if [ ! -z $pnet_pid ]; then
+            kill -15 $pnet_pid
+            echo "shutdown private network node"
+            rm .pnet    
+        fi
     fi
 
-    if [ ! -z $bootnode_pid ]; then
-        kill -9 $bootnode_pid
-        rm .bootnode
+    if [ -f .bootnode ]; then
+        local bootnode_pid=$(cat .bootnode)
+
+        if [ ! -z $bootnode_pid ]; then
+            kill -15 $bootnode_pid
+            echo "shutdown bootnode"
+            rm .bootnode
+        fi
     fi
 
-    if [ ! -z $node_pid ]; then
-        kill -9 $node_pid
-        rm .node
+    if [ -f .node ]; then
+        local node_pid=$(cat .node)
+
+        if [ ! -z $node_pid ]; then
+            kill -15 $node_pid
+            echo "shutdown node"
+            rm .node
+        fi
     fi
 }
 
 function add_node {
     if [ -f .node ]; then 
         node_pid=$(cat .node)
-        kill -9 $node_pid   
+        kill -15 $node_pid   
         echo "kill already running node process: $node_pid"
     fi
 
@@ -97,25 +111,26 @@ function add_node {
         geth --datadir ./var/node --port 30304 --networkid 333 --bootnodes "enode://26168c13637850b7b15c41a1f9cf1904cb93edde0c7479b0eaf6d018a4337842@localhost:30301" 2> ./var/log/node.log &
         node_pid=$! 
         echo $node_pid > .node
+        echo "start node"
     } || {
-        kill -9 $node_pid
+        kill -15 $node_pid
         rm .node
     }
 }
 
 function attach_pnet {
-    geth attach ipc:./var/main/geth.ipc
+    geth --datadir ./var/main attach ipc:./var/main/geth.ipc
 }
 
 case $1 in
-    clear) rm -rf ./var
-        exit 0
+    clear) kill_process
+        rm -rf ./var
         ;;
     attach) attach_pnet
         ;;
     ps) list_process
         ;;
-    kill) kill_process
+    shutdown) kill_process
         ;;
     add) add_node
         ;;
